requirements.txt
Flask==2.3.3
openai==1.3.5
Flask-Cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7 